// CaffStore.Frontend/src/app/api/generated/model/caffItemDtoPagedResponse.ts
/**
 * Caff Store API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CaffItemDto } from './caffItemDto';


export interface CaffItemDtoPagedResponse { 
    currentPage: number;
    pageSize: number;
    totalResultCount: number;
    readonly totalPageCount: number;
    results: Array<CaffItemDto>;
}